Factor out rapidpro_util module.
