+++ rapidpro/followup_campaigns.py
@@ -2,8 +2,10 @@ import datetime
 from typing import NamedTuple, Optional, List
 from django.conf import settings
 from temba_client.v2 import TembaClient
-from temba_client.v2.types import Group, Contact, Field
+from temba_client.v2.types import Contact
 from temba_client.utils import format_iso8601
+
+from .rapidpro_util import get_field, get_group, get_or_create_contact
 
 
 class DjangoSettingsFollowupCampaigns:
@@ -112,40 +114,3 @@ class FollowupCampaign(NamedTuple):
         if not value:
             return None
         return FollowupCampaign(*value.split(',', 1))
-
-
-def get_group(client: TembaClient, name: str) -> Group:
-    '''
-    Return the RapidPro group with the given name, raising an exception
-    if it doesn't exist.
-    '''
-
-    group = client.get_groups(name=name).first(retry_on_rate_exceed=True)
-    if group is None:
-        raise ValueError(f"Unable to find RapidPro group '{name}'")
-    return group
-
-
-def get_field(client: TembaClient, key: str) -> Field:
-    '''
-    Return the RapidPro field with the given key, raising an exception
-    if it doesn't exist.
-    '''
-
-    field = client.get_fields(key=key).first(retry_on_rate_exceed=True)
-    if field is None:
-        raise ValueError(f"Unable to find RapidPro field with key '{key}'")
-    return field
-
-
-def get_or_create_contact(client: TembaClient, name: str, phone_number: str) -> Contact:
-    '''
-    Retrieve the contact with the given phone number, creating them (and providing the
-    given name) if they don't already exist.
-    '''
-
-    urn = f'tel:+1{phone_number}'
-    contact = client.get_contacts(urn=urn).first(retry_on_rate_exceed=True)
-    if contact is None:
-        contact = client.create_contact(name=name, urns=[urn])
-    return contact
